# docs/api-docs.json
{"openapi":"3.1.0","info":{"title":"EMS APIs","description":"API documentation for Employee Management system","contact":{"name":"Rehan Sakiwala","email":"test@email.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/v1/employees/{id}":{"get":{"tags":["employee-controller"],"summary":"Get Employee By Id","description":"Fetches employee by unique id","operationId":"getEmployeeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Employee fetched successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDto"}}}},"404":{"description":"Employee not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDto"}}}},"500":{"description":"Server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDto"}}}}}},"put":{"tags":["employee-controller"],"summary":"Update employee","description":"Updates an entire employee record","operationId":"updateEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDto"}}}}}},"delete":{"tags":["employee-controller"],"summary":"Deleting Employee","operationId":"deleteemployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"patch":{"tags":["employee-controller"],"summary":"Partial update","description":"Updates specific fields of employee","operationId":"partialUpdateEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDto"}}}}}}},"/v1/employees":{"get":{"tags":["employee-controller"],"summary":"Get All Employees","description":"Get all employees","operationId":"getAllEmployees","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponseDto"}}}}}}},"post":{"tags":["employee-controller"],"summary":"Create Employee","description":"Creates a new employee","operationId":"createEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponseDto"}}}}}}},"/v1/employees/{id}/image":{"get":{"tags":["employee-controller"],"summary":"Get image","operationId":"getImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}},"post":{"tags":["employee-controller"],"summary":"Upload image","operationId":"uploadImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"EmployeeRequestDto":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string","minLength":1},"lastname":{"type":"string"},"password":{"type":"string"},"address":{"type":"string","minLength":1}},"required":["address","firstName"]},"EmployeeResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstname":{"type":"string"},"lastname":{"type":"string"},"address":{"type":"string"}}}}}}